+++ ktor/server/src/jvmMain/kotlin/dev/inmo/micro_utils/ktor/server/UniloadMultipart.kt
@@ -2,6 +2,7 @@ package dev.inmo.micro_utils.ktor.server
 
 import dev.inmo.micro_utils.common.*
 import dev.inmo.micro_utils.coroutines.safely
+import dev.inmo.micro_utils.ktor.common.downloadToTempFile
 import io.ktor.http.content.*
 import io.ktor.server.application.ApplicationCall
 import io.ktor.server.request.receiveMultipart
@@ -28,6 +29,7 @@ suspend inline fun ApplicationCall.handleUniUpload(
             is PartData.BinaryItem -> onBinaryContent(partData)
             is PartData.BinaryChannelItem -> onBinaryChannelItem(partData)
         }
+        partData.dispose()
     }
 }
 
@@ -36,7 +38,7 @@ suspend fun ApplicationCall.uniloadMultipart(
     onCustomFileItem: (PartData.FileItem) -> Unit = {},
     onBinaryChannelItem: (PartData.BinaryChannelItem) -> Unit = {},
     onBinaryContent: (PartData.BinaryItem) -> Unit = {}
-) = safely {
+): Input = safely {
     var resultInput: Input? = null
 
     handleUniUpload(
@@ -58,41 +60,11 @@ suspend fun ApplicationCall.uniloadMultipartFile(
     onCustomFileItem: (PartData.FileItem) -> Unit = {},
     onBinaryChannelItem: (PartData.BinaryChannelItem) -> Unit = {},
     onBinaryContent: (PartData.BinaryItem) -> Unit = {},
-) = safely {
-    val multipartData = receiveMultipart()
-
-    var resultInput: MPPFile? = null
-
-    multipartData.forEachPart {
-        when (it) {
-            is PartData.FormItem -> onFormItem(it)
-            is PartData.FileItem -> {
-                if (it.name == "bytes") {
-                    val name = FileName(it.originalFileName ?: error("File name is unknown for default part"))
-                    resultInput = MPPFile.createTempFile(
-                        name.nameWithoutExtension.let {
-                            var resultName = it
-                            while (resultName.length < 3) {
-                                resultName += "_"
-                            }
-                            resultName
-                        },
-                        ".${name.extension}"
-                    ).apply {
-                        outputStream().use { fileStream ->
-                            it.streamProvider().use {
-                                it.copyTo(fileStream)
-                            }
-                        }
-                    }
-                } else {
-                    onCustomFileItem(it)
-                }
-            }
-            is PartData.BinaryItem -> onBinaryContent(it)
-            is PartData.BinaryChannelItem -> onBinaryChannelItem(it)
-        }
-    }
-
-    resultInput ?: error("Bytes has not been received")
+): MPPFile = safely {
+    uniloadMultipart(
+        onFormItem,
+        onCustomFileItem,
+        onBinaryChannelItem,
+        onBinaryContent
+    ).downloadToTempFile()
 }